removing obsolete UserStore, adding UserOnlyStore w/ tests
