+++ src/ElCamino.AspNetCore.Identity.AzureTable/IdentityAzureTableBuilderExtensions.cs
@@ -9,28 +9,6 @@ namespace Microsoft.Extensions.DependencyInjection
 {
     public static class IdentityAzureTableBuilderExtensions
     {
-        [Obsolete("AddAzureTableStoresV2 will be renamed AddAzureTableStores in a future version. Use AddAzureTableStoresV2.")]
-        public static IdentityBuilder AddAzureTableStores<TContext>(this IdentityBuilder builder, Func<IdentityConfiguration> configAction)
-            where TContext : IdentityCloudContext, new()
-        {
-            builder.Services.AddSingleton<IdentityConfiguration>(new Func<IServiceProvider, IdentityConfiguration>(p => configAction()));
-
-            Type contextType = typeof(TContext);
-            Type userStoreType = typeof(UserStore<,,>).MakeGenericType(builder.UserType, builder.RoleType ?? typeof(ElCamino.AspNetCore.Identity.AzureTable.Model.IdentityRole), contextType);
-            Type roleStoreType = typeof(RoleStore<,>).MakeGenericType(builder.RoleType ?? typeof(ElCamino.AspNetCore.Identity.AzureTable.Model.IdentityRole), contextType);
-
-            builder.Services.AddScoped(contextType, contextType);
-
-            builder.Services.AddScoped(
-                typeof(IUserStore<>).MakeGenericType(builder.UserType),
-                userStoreType);
-            builder.Services.AddScoped(
-                typeof(IRoleStore<>).MakeGenericType(builder.RoleType ?? typeof(ElCamino.AspNetCore.Identity.AzureTable.Model.IdentityRole)),
-                roleStoreType);
-
-            return builder;
-        }
-
         public static IdentityBuilder AddAzureTableStoresV2<TContext>(this IdentityBuilder builder, Func<IdentityConfiguration> configAction)
             where TContext : IdentityCloudContext, new()
         {
