+++ src/ElCamino.AspNetCore.Identity.AzureTable/UserStoreV2.cs
@@ -15,205 +15,6 @@ using ElCamino.AspNetCore.Identity.AzureTable.Model;
 
 namespace ElCamino.AspNetCore.Identity.AzureTable
 {
-    /// <summary>
-    /// Supports IdentityUser with Roles, Claims, and Tokens as collection properties.
-    /// Use this for backwards compat existing code. Otherwise, data is the same.
-    /// </summary>
-    /// <typeparam name="TUser"></typeparam>
-    /// <typeparam name="TRole"></typeparam>
-    /// <typeparam name="TContext"></typeparam>
-    [Obsolete("UserStoreV2 will be renamed to UserStore in a future version. Use UserStoreV2 instead.")]
-    public class UserStore<TUser, TRole, TContext> : UserStoreV2<TUser, TRole, TContext>
-        , IUserStore<TUser>
-#pragma warning disable 0618
-        where TUser : Model.IdentityUser, new()
-#pragma warning restore 0618
-        where TRole : Model.IdentityRole<string, Model.IdentityUserRole>, new()
-        where TContext : IdentityCloudContext, new()
-    {
-        public UserStore(TContext context, IdentityConfiguration config) : base(context,config) { }
-
-        //Fixing code analysis issue CA1063
-        protected override void Dispose(bool disposing)
-        {
-            base.Dispose(disposing);
-        }
-
-        public override async Task<IList<TUser>> GetUsersForClaimAsync(Claim claim, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-            ThrowIfDisposed();
-
-            if (claim == null)
-            {
-                throw new ArgumentNullException(nameof(claim));
-            }
-
-
-            return (await GetUsersAggregateByIndexQueryAsync(GetUserByClaimQuery(claim), (userId) => {
-                return GetUserAggregateQueryAsync(userId, setFilterByUserId: null, whereRole: null, whereClaim: (uc) =>
-                {
-                    return uc.RowKey == KeyHelper.GenerateRowKeyIdentityUserClaim(claim.Type, claim.Value);
-                });
-
-            })).ToList();
-        }
-
-        public override async Task<IList<TUser>> GetUsersInRoleAsync(string roleName, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-            ThrowIfDisposed();
-
-            if (string.IsNullOrWhiteSpace(roleName))
-            {
-                throw new ArgumentException(IdentityResources.ValueCannotBeNullOrEmpty, nameof(roleName));
-            }
-
-            if (await RoleExistsAsync(roleName))
-            {
-                return (await this.GetUsersAggregateByIndexQueryAsync(GetUserByRoleQuery(roleName), (userId) => {
-                    return GetUserAggregateQueryAsync(userId, setFilterByUserId: null, whereClaim: null, whereRole: (ur) =>
-                    {
-                        return ur.RowKey == KeyHelper.GenerateRowKeyIdentityUserRole(roleName);
-                    });
-
-                })).ToList();
-            }
-
-            return new List<TUser>();
-        }
-
-        protected override async Task<TUser> GetUserAggregateAsync(TableQuery queryUser) 
-        {
-            var user = await _indexTable.ExecuteQueryAsync(queryUser).FirstOrDefaultAsync();
-            if (user != null)
-            {
-                string userId = user.Properties["Id"].StringValue;
-                var userResults = await GetUserAggregateQueryAsync(userId).ToListAsync();
-                var result =  MapUserAggregate(userId, userResults);
-                if(result.User != null)
-                {
-                    TUser u = result.User;
-                    foreach (var c in result.Claims)
-                    {
-                        u.Claims.Add(c);
-                    }
-                    foreach (var l in result.Logins)
-                    {
-                        u.Logins.Add(l);
-                    }
-                    foreach (var r in result.Roles)
-                    {
-                        u.Roles.Add(r);
-                    }
-                    foreach (var t in result.Tokens)
-                    {
-                        u.Tokens.Add(t);
-                    }
-                    return u;
-                }
-            }
-
-            return default(TUser);
-        }
-
-        private TUser MapUser(string userId, IEnumerable<DynamicTableEntity> userResults)
-        {
-            var result = MapUserAggregate(userId, userResults);
-            if (result.User != null)
-            {
-                TUser u = result.User;
-                foreach (var c in result.Claims)
-                {
-                    u.Claims.Add(c);
-                }
-                foreach (var l in result.Logins)
-                {
-                    u.Logins.Add(l);
-                }
-                foreach (var r in result.Roles)
-                {
-                    u.Roles.Add(r);
-                }
-                foreach (var t in result.Tokens)
-                {
-                    u.Tokens.Add(t);
-                }
-                return u;
-            }
-
-            return default(TUser);
-        }
-
-        protected override async Task<TUser> GetUserAsync(string userId)
-        {
-            return MapUser(userId, await GetUserAggregateQueryAsync(userId).ToListAsync());
-        }
-
-        protected async override Task<IEnumerable<TUser>> GetUserQueryAsync(IEnumerable<string> userIds)
-        {
-            const double pageSize = 50.0;
-            int pages = (int)Math.Ceiling(((double)userIds.Count() / pageSize));
-            List<TableQuery> listTqs = new List<TableQuery>(pages);
-            IEnumerable<string> tempUserIds = Enumerable.Empty<string>();
-
-            for (int currentPage = 1; currentPage <= pages; currentPage++)
-            {
-                if (currentPage > 1)
-                {
-                    tempUserIds = userIds.Skip(((currentPage - 1) * (int)pageSize)).Take((int)pageSize);
-                }
-                else
-                {
-                    tempUserIds = userIds.Take((int)pageSize);
-                }
-
-                TableQuery tq = new TableQuery();
-                int tempUserIdCounter = 0;
-                foreach (string tempUserId in tempUserIds)
-                {
-
-                    string temp = TableQuery.GenerateFilterCondition("PartitionKey", QueryComparisons.Equal, tempUserId);
-
-                    if (tempUserIdCounter > 0)
-                    {
-                        tq.FilterString = TableQuery.CombineFilters(tq.FilterString, TableOperators.Or, temp);
-                    }
-                    else
-                    {
-                        tq.FilterString = temp;
-                    }
-                    tempUserIdCounter++;
-                }
-                listTqs.Add(tq);
-
-            }
-
-            ConcurrentBag<TUser> bag = new ConcurrentBag<TUser>();
-#if DEBUG
-            DateTime startUserAggTotal = DateTime.UtcNow;
-#endif
-            var tasks = listTqs.Select((q) =>
-            {
-                return _userTable.ExecuteQueryAsync(q)
-                .ToListAsync()
-                .ContinueWith((urTask) => {
-                    foreach (var s in urTask.Result.GroupBy(g => g.PartitionKey))
-                    {
-                        bag.Add(MapUser(s.Key, s));
-                    }
-                });                                                                        
-            });
-            await Task.WhenAll(tasks).ConfigureAwait(false);
-#if DEBUG
-            Debug.WriteLine("GetUserAggregateQuery (GetUserAggregateTotal): {0} seconds", (DateTime.UtcNow - startUserAggTotal).TotalSeconds);
-#endif
-            return bag;
-        }
-
-    }
-
-    //-------------------------------------------------
     public class UserStoreV2<TUser, TContext> : UserStoreV2<TUser, Model.IdentityRole, string, Model.IdentityUserLogin, Model.IdentityUserRole, Model.IdentityUserClaim, Model.IdentityUserToken, TContext>
        , IUserStore<TUser>
        where TUser : Model.IdentityUser<string>, new()
