+++ tests/ElCamino.AspNetCore.Identity.AzureTable.Tests/Fixtures/RoleFixture.cs
@@ -6,15 +6,14 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using ElCamino.AspNetCore.Identity.AzureTable;
-using IdentityUser = ElCamino.AspNetCore.Identity.AzureTable.Model.IdentityUser;
 using IdentityRole = ElCamino.AspNetCore.Identity.AzureTable.Model.IdentityRole;
 using Microsoft.AspNetCore.Identity;
 
 namespace ElCamino.Web.Identity.AzureTable.Tests.Fixtures
 {
 #pragma warning disable 0618
-    public partial class RoleFixture<TUser, TRole, TContext> : BaseFixture<TUser, TRole, TContext, UserStore<TUser, TRole, TContext>>
-        where TUser : IdentityUser, new()
+    public partial class RoleFixture<TUser, TRole, TContext> : BaseFixture<TUser, TRole, TContext, UserStoreV2<TUser, TRole, TContext>>
+        where TUser : AspNetCore.Identity.AzureTable.Model.IdentityUserV2, new()
 #pragma warning restore 0618
         where TRole : IdentityRole, new()
         where TContext : IdentityCloudContext, new()
